// src/common/roles.enum.ts
export enum Role {
  Resident = 'resident',
  Representative = 'representative', // Representante de piso
  FloorAuditor = 'floor_auditor', // Fiscal de piso
  President = 'president', //presidente
  VicePresident = 'vice_president',
  GeneralAuditor = 'general_auditor', // Fiscal general
  Adjudicator = 'adjudicator', // Adjudicador
  SecretaryGeneral = 'secretary_general', // Secretario general
  Admin = 'admin',
}